refactor: Activity to Fragment
기능별로 activity를 ready, main, write 세가지 디렉토리에 나누었다.
ready는 activity로 구성해도 무관하다.
main은 각 페이지를 fragment로 개발해 MainActivity2.kt에 연결한다.
write는 각 페이지를 fragment로 개발해 WriteActivity.kt에 연결한다.

현재 MainActivity.kt의 'Fragment 구성으로 보기' 버튼에 MainActivity2.kt를 연결해 두었다.
