+++ app/src/main/java/com/nicer/attiary/ApplicationLifecycleObserver.kt
@@ -1,16 +1,8 @@
 package com.nicer.attiary
 
-import android.content.Intent
 import android.util.Log
-import androidx.core.app.ActivityCompat.startActivityForResult
 import androidx.lifecycle.DefaultLifecycleObserver
-import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.OnLifecycleEvent
-import com.nicer.attiary.data.app.AppLock
-import com.nicer.attiary.data.app.AppLockConst
-import com.nicer.attiary.data.app.lock
-import com.nicer.attiary.view.AppPassWordActivity
 
 class ApplicationLifecycleObserver : DefaultLifecycleObserver {
 	companion object {
