--- app/src/main/java/com/nicer/attiary/view/ready/AppPasswordFragment.kt
@@ -1,18 +1,0 @@
-package com.nicer.attiary.view.ready
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.nicer.attiary.R
-
-class AppPasswordFragment : Fragment() {
-
-	override fun onCreateView(
-		inflater: LayoutInflater, container: ViewGroup?,
-		savedInstanceState: Bundle?
-	): View? {
-		return inflater.inflate(R.layout.fragment_app_password, container, false)
-	}
-}
